backend/requirements.txt
fastapi[standard]
pydantic-settings
aiomysql
litellm
pyyaml
